# contracts/openapi/acr-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dev Health Agent Context Runtime API",
    "version": "0.1.0",
    "description": "Private hosted API for evidence-backed agent context packets, authorized provenance expansion, replay snapshots, and opt-in episode writeback. External Push is a separate API plane."
  },
  "servers": [
    {
      "url": "https://acr.fullchaos.dev"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/agent-context/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "x-required-scope": "context:read",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webAssertionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/ClientVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "Service, schema, entitlement, tool, and limit handshake",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/capabilities.v1.schema.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "426": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/agent-context/context-packets": {
      "post": {
        "operationId": "createContextPacket",
        "x-required-scope": "context:read",
        "x-required-entitlement": "agent_context_runtime",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webAssertionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/ClientVersion"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/context_packet_request.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated context packet; status may be complete, partial, degraded, or empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/context_packet.v1.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "426": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "504": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/agent-context/evidence/{evidence_ref_id}": {
      "get": {
        "operationId": "getEvidence",
        "x-required-scope": "evidence:read",
        "x-required-entitlement": "agent_context_runtime",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webAssertionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/ClientVersion"
          },
          {
            "name": "evidence_ref_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256
            }
          },
          {
            "name": "result_id",
            "in": "query",
            "required": false,
            "description": "Scopes an acr:v1 evidence reference to the stored investigation result of the answer being verified. Without it the newest citing result answers (deprecated, logged).",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expanded evidence after independent organization and repository authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/expanded_evidence.v1.schema.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "426": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "504": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/agent-context/episodes": {
      "post": {
        "operationId": "createEpisode",
        "x-required-scope": "episode:write",
        "x-required-entitlement": "agent_context_runtime",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/ClientVersion"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 256
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/agent_episode_create.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing episode returned for an identical idempotent retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/agent_episode.v1.schema.json"
                }
              }
            }
          },
          "201": {
            "description": "Episode created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/agent_episode.v1.schema.json"
                }
              }
            }
          },
          "204": {
            "description": "Episode accepted with no retained representation"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "426": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "504": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/oauth/device_authorization": {
      "post": {
        "security": [],
        "operationId": "createDeviceAuthorization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/device_authorization_request.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device authorization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/device_authorization_response.v1.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/oauth/token": {
      "post": {
        "security": [],
        "operationId": "exchangeDeviceToken",
        "description": "Dispatches by Content-Type (CHAOS-4013): application/json is the pre-existing RFC 8628 device-code grant, unchanged; application/x-www-form-urlencoded is the RFC 8693 workload token-exchange grant.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/device_token_request.v1.schema.json"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "grant_type",
                  "subject_token",
                  "subject_token_type"
                ],
                "properties": {
                  "grant_type": {
                    "const": "urn:ietf:params:oauth:grant-type:token-exchange"
                  },
                  "subject_token": {
                    "type": "string",
                    "description": "The k8s projected ServiceAccount JWT, validated via Kubernetes TokenReview."
                  },
                  "subject_token_type": {
                    "const": "urn:ietf:params:oauth:token-type:jwt"
                  },
                  "requested_token_type": {
                    "const": "urn:ietf:params:oauth:token-type:access_token"
                  },
                  "scope": {
                    "type": "string",
                    "description": "Space-delimited requested scope; may only narrow the resolved binding's grant, never widen it."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device grant redeemed once, or a workload token successfully exchanged",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "../jsonschema/v1/device_token_response.v1.schema.json"
                    },
                    {
                      "$ref": "../jsonschema/v1/token_exchange_response.v1.schema.json"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Pinned RFC device grant error, or an RFC 8693 token-exchange error",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "../jsonschema/v1/oauth_device_error.v1.schema.json"
                    },
                    {
                      "$ref": "../jsonschema/v1/oauth_token_exchange_error.v1.schema.json"
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/oauth/device_approval": {
      "post": {
        "security": [
          {
            "webAssertionAuth": []
          }
        ],
        "operationId": "approveDeviceAuthorization",
        "x-required-permission": "credential:issue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "../jsonschema/v1/device_approval_request.v1.schema.json"
                  },
                  {
                    "$ref": "../jsonschema/v1/device_approval_preview_request.v1.schema.json"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bounded device grant approved or advisory hints previewed",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "../jsonschema/v1/device_approval_response.v1.schema.json"
                    },
                    {
                      "$ref": "../jsonschema/v1/device_approval_preview_response.v1.schema.json"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/credentials/self/rotate": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "rotateOwnCredential",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/credential_rotate_request.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current credential rotated with rollback receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/credential_rotate_response.v1.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/credentials/self/revoke": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "revokeOwnCredential",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../jsonschema/v1/credential_revoke_request.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current credential revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../jsonschema/v1/credential_revoke_response.v1.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fcacr_ client credential"
      },
      "webAssertionAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ACR-Web-Assertion",
        "description": "Ed25519 request-bound compact JWT assertion for read-only server-to-server web requests. Assertions require the configured issuer, audience, local JWKS key ID, exact method/path/body hash, explicit repository scopes, and a maximum 30-second lifetime."
      }
    },
    "parameters": {
      "RequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional correlation ID for request tracing and audit. The service generates and returns a request ID when absent or invalid. It does not authenticate or authorize the caller.",
        "schema": {
          "type": "string",
          "maxLength": 256
        }
      },
      "ClientVersion": {
        "name": "X-ACR-Client-Version",
        "in": "header",
        "required": true,
        "description": "Required SemVer compatibility signal for protected routes. It is advisory only and does not authenticate or authorize the caller; unsupported versions return 426 version_mismatch.",
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Structured safe error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../jsonschema/v1/error.v1.schema.json"
            }
          }
        }
      }
    }
  }
}